+++ web/main.py
@@ -31,7 +31,7 @@ from app.brushtask import BrushTask
 from app.conf import ModuleConf, SystemConfig
 from app.downloader import Downloader
 from app.filter import Filter
-from app.helper import SecurityHelper, MetaHelper, ChromeHelper, ThreadHelper
+from app.helper import SecurityHelper, MetaHelper, ChromeHelper, ThreadHelper, CleanHelper
 from app.indexer import Indexer
 from app.media.meta import MetaInfo
 from app.mediaserver import MediaServer
@@ -776,10 +776,13 @@ def service():
         if not SystemUtils.is_docker() or not SystemUtils.get_all_processes():
             Services.pop('processes')
 
+    CleanConf = CleanHelper.get_default_config()
     return render_template("service.html",
                            Count=len(Services),
                            RuleGroups=RuleGroups,
                            SyncPaths=SyncPaths,
+                           CleanDefaultRoot=CleanConf.get("root_path"),
+                           CleanDefaultThreshold=CleanConf.get("threshold_mb"),
                            SchedulerTasks=Services)